main: extract run lifecycle domain from state.ts
Stage D of the state.ts decomposition: startRun, confirmRun,
rejectRun, cancelRun, driveRun, run snapshots, and run log/step
helpers move verbatim into runs.ts behind an explicit RunServiceHost
interface; RunDeliveryService reaches the log/step helpers through the
run service. confirmRun verified byte-identical after host-call
normalization; run-confirmation and write-plan invariant tests pass
unchanged. state.ts drops from 5,247 to 4,678 lines.
