+++ apps/desktop/electron/state.ts
@@ -14,13 +14,9 @@ import {
   createMsalClient,
   createOllamaLlm,
   createRegistryInstallCountPayload,
-  createQueuedRun,
   createTenantSession,
   DEFAULT_SCOPE_METADATA,
   probeSubscribedSkus,
-  executeApply,
-  executePlan,
-  executeRun,
   findConnectorFactory,
   findRegistryAgentById,
   listAllRegistryAgents,
@@ -33,7 +29,6 @@ import {
   removeAccount,
   runInteractiveFlow,
   setAgentUpdatesDir,
-  tenantSatisfiesRequirement,
   toInstalledAgent,
   type TokenCacheStorage,
 } from "@openadminos/runtime";
@@ -59,7 +54,6 @@ import type {
   RunHistoryPruneResult,
   RunHistoryPruneTrigger,
   RunHistoryRetentionSettings,
-  RunStepStatus,
   ProviderTestResult,
   StartRunOptions,
   TenantRecord,
@@ -111,7 +105,6 @@ import {
   DEFAULT_RUN_HISTORY_RETENTION_SETTINGS,
   providerCatalog,
   resolveProviderDefaultModel,
-  resolveRunModel,
   type AgentDiscordDelivery,
   type AgentOutlookDelivery,
   type AgentSchedule,
@@ -139,7 +132,6 @@ import {
 import { SafeStorageTokenCacheStore } from "./secret-store.js";
 import { SafeStorageConnectorSecretStore } from "./connector-secret-store.js";
 import { SafeStorageProviderSecretStore } from "./provider-secret-store.js";
-import { requestConnectorConfirmation } from "./connector-confirm-bridge.js";
 import { searchEndpoints } from "./graph-catalog.js";
 
 import {
@@ -149,13 +141,13 @@ import {
 } from "./registry-client.js";
 import { IntelligenceSqliteStore } from "./intune-chat/sqlite-store.js";
 import { IntuneChatService } from "./intune-chat/service.js";
+import { RunService } from "./runs.js";
 import { RunDeliveryService } from "./run-delivery.js";
 import {
   OUTLOOK_CONNECTOR_ID,
   SLACK_CONNECTOR_ID,
   DISCORD_CONNECTOR_ID,
   SIGNAL_CONNECTOR_ID,
-  fingerprintRunOutput,
   whatsappWebStatusToConnectorStatus,
   type RunDeliveryQueueItem,
 } from "./run-delivery-format.js";
@@ -195,11 +187,9 @@ import {
   humanizeMsalError,
   humanizeScheduledRunError,
   isNodeError,
-  isTerminalRunStatus,
   normalizeWorkspaceInstructions,
   normalizeWorkspaceTitle,
   tenantNamesById,
-  withSelfTrainingOverlay,
 } from "./state-helpers.js";
 export { __agentDraftTestUtils } from "./agent-draft-helpers.js";
 
@@ -949,6 +939,7 @@ export class AppStateStore {
   private readonly userDataPath: string | undefined;
   private readonly intelligenceStore: IntelligenceSqliteStore | undefined;
   private readonly chatService: IntuneChatService;
+  private readonly runService: RunService;
   private readonly runDeliveryService: RunDeliveryService;
   private readonly graphFactory: AppStateStoreOptions["graphFactory"] | undefined;
   private readonly llmFactory: AppStateStoreOptions["llmFactory"] | undefined;
@@ -965,11 +956,6 @@ export class AppStateStore {
   private readonly providerSecretStore: SafeStorageProviderSecretStore | undefined;
   private readonly allowDevRegistrySource: boolean;
   private msalClient: PublicClientApplication | undefined;
-  // Soft-cancel set. While a run id is here, progress snapshots from
-  // the runtime are dropped so the run stays in "cancelled" state. The
-  // background driver eventually returns; we don't (yet) plumb an
-  // AbortSignal through the runtime to interrupt it mid-flight.
-  private readonly cancelledRunIds = new Set<string>();
   private whatsappWebClientInstance: WhatsAppWebClientLike | undefined;
 
   // Registry cache — populated by initRegistry(), falls back to
@@ -1056,6 +1042,29 @@ export class AppStateStore {
         return host.graphFactory;
       },
     });
+    this.runService = new RunService({
+      read: () => host.read(),
+      write: (state) => host.write(state as PersistedState),
+      serialize: (task) => host.serialize(task),
+      listProviders: () => host.listProviders(),
+      providerCanRun: (provider) => host.providerCanRun(provider),
+      buildLlm: (providerId, model) => host.buildLlm(providerId, model),
+      buildGraph: (pinnedTenantId, agentScopes) =>
+        host.buildGraph(pinnedTenantId, agentScopes),
+      readConnectorConfigs: () => host.readConnectorConfigs(),
+      connectorSecretsFor: (connectorId) => host.connectorSecretsFor(connectorId),
+      selfTrainingPromptOverlay: (tenantId, agentSlug) =>
+        host.selfTrainingPromptOverlay(tenantId, agentSlug),
+      recordLearningEventSafely: (input) => host.recordLearningEventSafely(input),
+      notifyRunFinished: (run) => host.notifyRunFinished(run),
+      emitStateChanged: (reason, runId) => host.emitStateChanged(reason, runId),
+      enqueueRunDeliveries: (run) => host.enqueueRunDeliveries(run),
+      processPendingRunDeliveries: () => host.processPendingRunDeliveries(),
+      appVersion: this.appVersion,
+      get intelligenceStore() {
+        return host.intelligenceStore;
+      },
+    });
     this.runDeliveryService = new RunDeliveryService({
       read: () => host.read(),
       write: (state) => host.write(state as PersistedState),
@@ -1066,10 +1075,10 @@ export class AppStateStore {
       connectorSecretsFor: (connectorId) => host.connectorSecretsFor(connectorId),
       whatsAppWebClient: () => host.whatsAppWebClient(),
       appendRunLog: (runId, level, message, metadata) =>
-        host.appendRunLog(runId, level, message, metadata),
-      appendRunStep: (runId, input) => host.appendRunStep(runId, input),
+        host.runService.appendRunLog(runId, level, message, metadata),
+      appendRunStep: (runId, input) => host.runService.appendRunStep(runId, input),
       finishRunStep: (runId, stepId, status, detail) =>
-        host.finishRunStep(runId, stepId, status, detail),
+        host.runService.finishRunStep(runId, stepId, status, detail),
     });
     // Point the runtime at the OTA-updated manifest tree (if userData is
     // configured). When an agent has been updated via `updateAgent`, the
@@ -3468,61 +3477,7 @@ Return ONLY the YAML manifest. Do not include any commentary, headings, or markd
   }
 
   async cancelRun(runId: string): Promise<RunRecord> {
-    const result = await this.serialize(async () => {
-      const persisted = await this.read();
-      const run = persisted.runs.find((existing) => existing.id === runId);
-      if (!run) {
-        throw new Error(`Run not found: ${runId}`);
-      }
-      if (
-        run.status === "completed" ||
-        run.status === "failed" ||
-        run.status === "rejected" ||
-        run.status === "cancelled"
-      ) {
-        // Already terminal — nothing to cancel.
-        return run;
-      }
-
-      const finishedAt = new Date().toISOString();
-      const cancelled: RunRecord = {
-        ...run,
-        status: "cancelled",
-        finishedAt,
-        // Overwrite any stale in-progress summary (e.g. "X is running.")
-        // with an explicit cancellation message. The original summary
-        // would otherwise leak into Activity rows long after the cancel.
-        summary: "Cancelled by user.",
-        // Transition any in-flight step to "cancelled" and stop any
-        // streaming reasoning indicator. Without this the UI keeps
-        // spinning the active step and showing a "streaming" badge
-        // even though the run is terminal.
-        steps: run.steps.map((step) =>
-          step.status === "running"
-            ? {
-                ...step,
-                status: "cancelled",
-                finishedAt: step.finishedAt ?? finishedAt,
-                thinking: step.thinking
-                  ? { ...step.thinking, streaming: false }
-                  : step.thinking,
-              }
-            : step.thinking?.streaming
-              ? {
-                  ...step,
-                  thinking: { ...step.thinking, streaming: false },
-                }
-              : step,
-        ),
-      };
-      const nextRuns = persisted.runs.map((existing) =>
-        existing.id === runId ? cancelled : existing,
-      );
-      await this.write({ ...persisted, runs: nextRuns });
-      return cancelled;
-    });
-    this.cancelledRunIds.add(runId);
-    return result;
+    return this.runService.cancelRun(runId);
   }
 
   async installAgent(agentId: string): Promise<AppState> {
@@ -4206,312 +4161,19 @@ Return ONLY the YAML manifest. Do not include any commentary, headings, or markd
     agentSlug: string,
     options: StartRunOptions = {},
   ): Promise<RunRecord> {
-    const queued = await this.serialize(async () => {
-      const persisted = await this.read();
-      const agent = persisted.installedAgents.find(
-        (installedAgent) => installedAgent.slug === agentSlug,
-      );
-
-      if (!agent) {
-        throw new Error(`Agent is not installed: ${agentSlug}`);
-      }
-      assertAgentCompatible(
-        withAgentCompatibility(agent, this.appVersion),
-        "run",
-      );
-
-      const providers = await this.listProviders();
-      // Honor a per-run provider override if supplied; otherwise fall
-      // back to the globally-active provider. Unknown ids are an error
-      // — silently dropping the override would be misleading.
-      let selectedProvider: ProviderSummary | undefined;
-      if (options.providerId !== undefined) {
-        selectedProvider = providers.find((p) => p.id === options.providerId);
-        if (!selectedProvider) {
-          throw new Error(`Unknown provider: ${String(options.providerId)}`);
-        }
-      } else {
-        selectedProvider =
-          providers.find((provider) => provider.id === persisted.activeProviderId) ??
-          providers[0];
-      }
-      const activeProvider = selectedProvider;
-      const providerId =
-        activeProvider?.id ?? options.providerId ?? persisted.activeProviderId;
-
-      // Resolve which model to stamp on the run, in priority order:
-      //   1. Explicit per-run override (options.model) when supplied
-      //   2. Agent manifest's preferredModel IF the provider has it pulled
-      //   3. User's pinned activeModelByProviderId[providerId] if set
-      //   4. Provider's first reported model (defaultModel)
-      const knownModels = activeProvider?.models ?? [];
-      const explicitModel =
-        typeof options.model === "string" && options.model.length > 0
-          ? options.model
-          : undefined;
-      if (explicitModel) {
-        if (knownModels.length > 0 && !knownModels.includes(explicitModel)) {
-          const recovery =
-            activeProvider?.id === "ollama"
-              ? ` Pull it with \`ollama pull ${explicitModel}\` and try again.`
-              : " Pick one of the models reported by the provider and try again.";
-          throw new Error(
-            `Model "${explicitModel}" is not available for ${activeProvider?.name ?? providerId}.${recovery}`,
-          );
-        }
-      }
-      const model = resolveRunModel({
-        provider: activeProvider,
-        activeModelByProviderId: persisted.activeModelByProviderId,
-        preferredModel: agent.preferredModel,
-        explicitModel,
-      }).model;
-
-      // Preflight the LLM provider so a clearly-actionable error is
-      // returned to the renderer synchronously instead of a queued
-      // run that fails moments later when the runtime can't reach it.
-      if (activeProvider && !this.providerCanRun(activeProvider)) {
-        if (activeProvider.id === "ollama") {
-          throw new Error(
-            "Ollama isn't reachable. Start it with `ollama serve`, then try again.",
-          );
-        }
-        throw new Error(
-          `${activeProvider.name} isn't ready. Open Settings → LLM Providers to check the connection.`,
-        );
-      }
-
-      // Resolve the effective tenant at queue time. Runs cannot proceed
-      // without a connected tenant — onboarding is the gate that gets a
-      // user here in the first place, but defend in depth.
-      //   - explicit id  -> validate it exists and pin it
-      //   - omitted      -> default to currently-active tenant
-      let pinnedTenantId: string;
-      if (typeof options.tenantId === "string") {
-        const exists = persisted.tenants.some((tenant) => tenant.id === options.tenantId);
-        if (!exists) {
-          throw new Error(`Tenant not connected: ${options.tenantId}`);
-        }
-        pinnedTenantId = options.tenantId;
-      } else if (persisted.activeTenantId) {
-        pinnedTenantId = persisted.activeTenantId;
-      } else {
-        throw new Error(
-          "No tenant connected. Connect a Microsoft 365 tenant before running agents.",
-        );
-      }
-
-      // Entra ID tier preflight: if the agent declares a required tier
-      // and the tenant's detected tier is known to fall short, refuse
-      // the run with a clear remediation message. `unknown` tier (not
-      // probed yet, or probe failed) is treated as informational —
-      // runs proceed and the actual Graph call may fail with a real
-      // 403, which still surfaces meaningfully via the runtime.
-      const requiredTier = agent.requiresEntraTier ?? "free";
-      if (requiredTier !== "free") {
-        const tenantRecord = persisted.tenants.find((t) => t.id === pinnedTenantId);
-        const satisfies = tenantSatisfiesRequirement(tenantRecord?.entraTier, requiredTier);
-        if (satisfies === false) {
-          const detectedLabel = tenantRecord?.entraTier === "free" ? "Entra ID Free" : `Entra ID ${tenantRecord?.entraTier?.toUpperCase()}`;
-          const requiredLabel = `Entra ID ${requiredTier.toUpperCase()}`;
-          throw new Error(
-            `${agent.name} requires ${requiredLabel}. The active tenant (${tenantRecord?.displayName ?? pinnedTenantId}) is on ${detectedLabel}. Microsoft 365 Business Premium includes Entra ID P1 — check your tenant's subscription, or pick a free-tier agent.`,
-          );
-        }
-      }
-
-      const queuedRun = createQueuedRun({ agent, providerId, model });
-      queuedRun.tenantId = pinnedTenantId;
-      queuedRun.trigger = options.trigger ?? "manual";
-
-      await this.write({
-        ...persisted,
-        runs: [queuedRun, ...persisted.runs],
-      });
-
-      return { agent, providerId, model, queuedRun };
-    });
-
-    if (options.source?.type === "intune-chat" && this.intelligenceStore) {
-      const now = new Date().toISOString();
-      const statusNote =
-        queued.agent.mode === "write"
-          ? "Write actions still require the normal confirmation flow."
-          : "Read-only agent run queued.";
-      const message: IntuneChatMessage = {
-        id: `msg_${randomUUID()}`,
-        conversationId: options.source.conversationId,
-        role: "assistant",
-        content: `${queued.agent.name} started from chat. Run ${queued.queuedRun.id} is ${queued.queuedRun.status}. ${statusNote}`,
-        status: "completed",
-        providerId: queued.providerId,
-        model: queued.model,
-        createdAt: now,
-      };
-      this.intelligenceStore.insertMessage(message);
-      this.intelligenceStore.insertToolCall({
-        id: `tool_${randomUUID()}`,
-        conversationId: options.source.conversationId,
-        messageId: message.id,
-        type: "agent-run",
-        status: "completed",
-        createdAt: now,
-        completedAt: now,
-        input: {
-          agentSlug,
-          originatingMessageId: options.source.messageId,
-        },
-        output: {
-          runId: queued.queuedRun.id,
-          runStatus: queued.queuedRun.status,
-        },
-      });
-      this.intelligenceStore.touchConversation(options.source.conversationId, undefined, now);
-    }
-
-    void this.driveRun({
-      run: queued.queuedRun,
-      agent: queued.agent,
-      providerId: queued.providerId,
-      model: queued.model,
-    });
-    this.recordLearningEventSafely({
-      tenantId: queued.queuedRun.tenantId,
-      agentSlug: queued.agent.slug,
-      eventType: "agent.run-started",
-      source: "run",
-      payload: {
-        runId: queued.queuedRun.id,
-        trigger: queued.queuedRun.trigger,
-        mode: queued.agent.mode,
-      },
-    });
-    return queued.queuedRun;
+    return this.runService.startRun(agentSlug, options);
   }
 
   async confirmRun(runId: string, phrase: string): Promise<RunRecord> {
-    const transition = await this.serialize(async () => {
-      const persisted = await this.read();
-      const run = persisted.runs.find((existing) => existing.id === runId);
-      if (!run) {
-        throw new Error(`Run not found: ${runId}`);
-      }
-      if (run.status !== "awaiting-confirmation") {
-        throw new Error(
-          `Run ${runId} is not awaiting confirmation (status: ${run.status}).`,
-        );
-      }
-      if (!run.plan) {
-        throw new Error(`Run ${runId} has no plan to confirm.`);
-      }
-      if (phrase !== run.plan.confirmationPhrase) {
-        throw new Error("Confirmation phrase does not match.");
-      }
-
-      const agent = persisted.installedAgents.find(
-        (installedAgent) => installedAgent.slug === run.agentSlug,
-      );
-      if (!agent) {
-        throw new Error(`Agent is not installed: ${run.agentSlug}`);
-      }
-      if (agent.mode !== "write") {
-        throw new Error(`Agent ${run.agentSlug} is not a write agent.`);
-      }
-
-      const confirmedAt = new Date().toISOString();
-      const updated: RunRecord = {
-        ...run,
-        status: "running",
-        confirmedAt,
-        startedAt: confirmedAt,
-        summary: `${agent.name} is applying.`,
-      };
-      await this.write({
-        ...persisted,
-        runs: persisted.runs.map((existing) =>
-          existing.id === runId ? updated : existing,
-        ),
-      });
-
-      const providers = await this.listProviders();
-      const activeProvider =
-        providers.find((provider) => provider.id === persisted.activeProviderId) ??
-        providers[0];
-      const providerId = run.providerId ?? activeProvider?.id ?? persisted.activeProviderId;
-      const model =
-        run.model ??
-        resolveProviderDefaultModel(
-          activeProvider,
-          persisted.activeModelByProviderId,
-        ).model;
-
-      return { agent, providerId, model, updated };
-    });
-
-    void this.driveApply({
-      run: transition.updated,
-      agent: transition.agent,
-      providerId: transition.providerId,
-      model: transition.model,
-      plan: transition.updated.plan!,
-    });
-    this.recordLearningEventSafely({
-      tenantId: transition.updated.tenantId,
-      agentSlug: transition.agent.slug,
-      eventType: "agent.write-plan-confirmed",
-      source: "run",
-      payload: {
-        runId: transition.updated.id,
-        actionCount: transition.updated.plan?.actions.length ?? 0,
-      },
-    });
-    return transition.updated;
+    return this.runService.confirmRun(runId, phrase);
   }
 
   async rejectRun(runId: string): Promise<RunRecord> {
-    return this.serialize(async () => {
-      const persisted = await this.read();
-      const run = persisted.runs.find((existing) => existing.id === runId);
-      if (!run) {
-        throw new Error(`Run not found: ${runId}`);
-      }
-      if (run.status !== "awaiting-confirmation") {
-        throw new Error(
-          `Run ${runId} cannot be rejected (status: ${run.status}).`,
-        );
-      }
-
-      const rejectedAt = new Date().toISOString();
-      const updated: RunRecord = {
-        ...run,
-        status: "rejected",
-        rejectedAt,
-        finishedAt: rejectedAt,
-        summary: `Run rejected by user.`,
-      };
-      await this.write({
-        ...persisted,
-        runs: persisted.runs.map((existing) =>
-          existing.id === runId ? updated : existing,
-        ),
-      });
-      this.recordLearningEventSafely({
-        tenantId: updated.tenantId,
-        agentSlug: updated.agentSlug,
-        eventType: "agent.write-plan-rejected",
-        source: "run",
-        payload: {
-          runId: updated.id,
-          actionCount: updated.plan?.actions.length ?? 0,
-        },
-      });
-      return updated;
-    });
+    return this.runService.rejectRun(runId);
   }
 
   async getRun(id: string): Promise<RunRecord | undefined> {
-    const persisted = await this.read();
-    return persisted.runs.find((run) => run.id === id);
+    return this.runService.getRun(id);
   }
 
   private async buildLlm(
@@ -4661,130 +4323,8 @@ Return ONLY the YAML manifest. Do not include any commentary, headings, or markd
     };
   }
 
-  private stampTenant(run: RunRecord, tenantId: string): RunRecord {
-    return { ...run, tenantId };
-  }
-
-  private async driveRun(input: {
-    run: RunRecord;
-    agent: AgentSummary;
-    providerId: ProviderId;
-    model?: string;
-  }): Promise<void> {
-    try {
-      const driver = input.agent.mode === "write" ? executePlan : executeRun;
-      const baseLlm = await this.buildLlm(input.providerId, input.model);
-      const selection = await this.buildGraph(input.run.tenantId, input.agent.scopes);
-      const overlay = this.selfTrainingPromptOverlay(selection.tenantId, input.agent.slug);
-      const llm = overlay ? withSelfTrainingOverlay(baseLlm, overlay) : baseLlm;
-      const stampedRun = this.stampTenant(input.run, selection.tenantId);
-      await this.persistRunSnapshot(stampedRun);
-      await driver({
-        run: stampedRun,
-        agent: input.agent,
-        providerId: input.providerId,
-        model: input.model,
-        llm,
-        createGraph: selection.createGraph,
-        tenant: selection.tenantSession,
-        connectorConfigs: await this.readConnectorConfigs(),
-        connectorSecretsFor: (connectorId) => this.connectorSecretsFor(connectorId),
-        confirmCapability: requestConnectorConfirmation,
-        realWrites: true,
-        onProgress: (next) =>
-          this.persistRunSnapshot(this.stampTenant(next, selection.tenantId)),
-      });
-    } catch (error) {
-      await this.persistFailedSnapshot(input.run, input.agent, error);
-    }
-  }
-
-  private async driveApply(input: {
-    run: RunRecord;
-    agent: AgentSummary;
-    providerId: ProviderId;
-    model?: string;
-    plan: NonNullable<RunRecord["plan"]>;
-  }): Promise<void> {
-    try {
-      const baseLlm = await this.buildLlm(input.providerId, input.model);
-      const selection = await this.buildGraph(input.run.tenantId, input.agent.scopes);
-      const overlay = this.selfTrainingPromptOverlay(selection.tenantId, input.agent.slug);
-      const llm = overlay ? withSelfTrainingOverlay(baseLlm, overlay) : baseLlm;
-      await executeApply({
-        run: input.run,
-        agent: input.agent,
-        providerId: input.providerId,
-        model: input.model,
-        plan: input.plan,
-        llm,
-        createGraph: selection.createGraph,
-        tenant: selection.tenantSession,
-        connectorConfigs: await this.readConnectorConfigs(),
-        connectorSecretsFor: (connectorId) => this.connectorSecretsFor(connectorId),
-        confirmCapability: requestConnectorConfirmation,
-        realWrites: true,
-        onProgress: (next) =>
-          this.persistRunSnapshot(this.stampTenant(next, selection.tenantId)),
-      });
-    } catch (error) {
-      await this.persistFailedSnapshot(input.run, input.agent, error);
-    }
-  }
-
-  private async persistFailedSnapshot(
-    run: RunRecord,
-    agent: AgentSummary,
-    error: unknown,
-  ): Promise<void> {
-    const message = error instanceof Error ? error.message : String(error);
-    const finishedAt = new Date().toISOString();
-    await this.persistRunSnapshot({
-      ...run,
-      status: "failed",
-      finishedAt,
-      summary: `${agent.name} failed: ${message}`,
-      error: message,
-    });
-  }
-
   private async persistRunSnapshot(run: RunRecord): Promise<void> {
-    if (this.cancelledRunIds.has(run.id)) {
-      // Run was soft-cancelled: discard further progress snapshots so
-      // the stored state stays in the "cancelled" terminal state even
-      // while background work finishes returning.
-      return Promise.resolve();
-    }
-    let deliveryCandidate: RunRecord | undefined;
-    await this.serialize(async () => {
-      const persisted = await this.read();
-      const previous = persisted.runs.find((existing) => existing.id === run.id);
-      const wasTerminal = previous ? isTerminalRunStatus(previous.status) : false;
-      const isNowTerminal = isTerminalRunStatus(run.status);
-      const nextRun =
-        isNowTerminal && run.status === "completed" && run.trigger === "schedule"
-          ? this.withScheduleChangeState(run, persisted.runs)
-          : run;
-      const exists = previous !== undefined;
-      const nextRuns = exists
-        ? persisted.runs.map((existing) => (existing.id === nextRun.id ? nextRun : existing))
-        : [nextRun, ...persisted.runs];
-      await this.write({ ...persisted, runs: nextRuns });
-      if (!wasTerminal && isNowTerminal && this.onRunFinished) {
-        try {
-          this.onRunFinished(nextRun);
-        } catch (error) {
-          console.error("[state] onRunFinished listener failed", error);
-        }
-      }
-      if (!wasTerminal && isNowTerminal) {
-        deliveryCandidate = nextRun;
-      }
-    });
-    if (deliveryCandidate) {
-      await this.enqueueRunDeliveries(deliveryCandidate);
-      void this.processPendingRunDeliveries();
-    }
+    return this.runService.persistRunSnapshot(run);
   }
 
   async processPendingRunDeliveries(): Promise<void> {
@@ -4795,122 +4335,13 @@ Return ONLY the YAML manifest. Do not include any commentary, headings, or markd
     return this.runDeliveryService.enqueueRunDeliveries(run);
   }
 
-  private withScheduleChangeState(run: RunRecord, runs: RunRecord[]): RunRecord {
-    const previous = runs.find(
-      (candidate) =>
-        candidate.id !== run.id &&
-        candidate.agentSlug === run.agentSlug &&
-        candidate.trigger === "schedule" &&
-        candidate.status === "completed",
-    );
-    if (!previous) return { ...run, changeState: "new" };
-    return {
-      ...run,
-      changeState:
-        fingerprintRunOutput(previous) === fingerprintRunOutput(run)
-          ? "unchanged"
-          : "changed",
-    };
-  }
-
-  private async appendRunLog(
-    runId: string,
-    level: RunLogLevel,
-    message: string,
-    metadata?: Record<string, unknown>,
-  ): Promise<void> {
-    await this.serialize(async () => {
-      const persisted = await this.read();
-      const timestamp = new Date().toISOString();
-      const nextRuns = persisted.runs.map((run) =>
-        run.id === runId
-          ? {
-              ...run,
-              logs: [
-                ...run.logs,
-                {
-                  id: `log_${randomUUID()}`,
-                  runId,
-                  timestamp,
-                  level,
-                  message,
-                  ...(metadata ? { metadata } : {}),
-                },
-              ],
-            }
-          : run,
-      );
-      await this.write({ ...persisted, runs: nextRuns });
-    });
-    this.emitStateChanged("run-log-appended", runId);
-  }
-
-  private async appendRunStep(
-    runId: string,
-    input: {
-      label: string;
-      status: RunStepStatus;
-      detail?: string;
-    },
-  ): Promise<string> {
-    const stepId = `step_delivery_${randomUUID().slice(0, 8)}`;
-    const timestamp = new Date().toISOString();
-    await this.serialize(async () => {
-      const persisted = await this.read();
-      const nextRuns = persisted.runs.map((run) =>
-        run.id === runId
-          ? {
-              ...run,
-              steps: [
-                ...run.steps,
-                {
-                  id: stepId,
-                  runId,
-                  label: input.label,
-                  status: input.status,
-                  ...(input.detail ? { detail: input.detail } : {}),
-                  startedAt: timestamp,
-                  ...(input.status === "running" ? {} : { finishedAt: timestamp }),
-                },
-              ],
-            }
-          : run,
-      );
-      await this.write({ ...persisted, runs: nextRuns });
-    });
-    this.emitStateChanged("run-step-appended", runId);
-    return stepId;
-  }
-
-  private async finishRunStep(
-    runId: string,
-    stepId: string,
-    status: Extract<RunStepStatus, "completed" | "failed" | "skipped">,
-    detail?: string,
-  ): Promise<void> {
-    const finishedAt = new Date().toISOString();
-    await this.serialize(async () => {
-      const persisted = await this.read();
-      const nextRuns = persisted.runs.map((run) =>
-        run.id === runId
-          ? {
-              ...run,
-              steps: run.steps.map((step) =>
-                step.id === stepId
-                  ? {
-                      ...step,
-                      status,
-                      finishedAt,
-                      ...(detail ? { detail } : {}),
-                    }
-                  : step,
-              ),
-            }
-          : run,
-      );
-      await this.write({ ...persisted, runs: nextRuns });
-    });
-    this.emitStateChanged("run-step-updated", runId);
+  private notifyRunFinished(run: RunRecord): void {
+    if (!this.onRunFinished) return;
+    try {
+      this.onRunFinished(run);
+    } catch (error) {
+      console.error("[state] onRunFinished listener failed", error);
+    }
   }
 
   private emitStateChanged(reason: string, runId?: string): void {
